wip: extract fixtures
